feat: swap old app with new one with navigation
